+++ src/pages/product/category/ProductCategoryPage.tsx
@@ -3,7 +3,7 @@ import { useFetchProductCategory } from '@shared/hooks'
 import { ProductCardSection } from '@widgets/product/product-home/layout'
 import {
   CategoryMenu,
-  ProductAddButton,
+  ProductCreateButton,
   Ranking,
 } from '@widgets/product/product-home/components'
 
@@ -12,7 +12,6 @@ import { useProductStore } from '@/src/shared/store/productStore'
 import useFetchProductsByQuery from '@/src/shared/hooks/query/product.query'
 
 import { NextSeo } from 'next-seo'
-
 
 interface ProductCategoryPageProps {
   category: Categories
@@ -24,7 +23,6 @@ export default function ProductCategoryPage({
   const { inputValue } = useProductStore()
   const { data } = useFetchProductsByQuery(category.id, inputValue)
   const productList = data!
-  console.log(data)
 
   return (
     <>
@@ -43,7 +41,7 @@ export default function ProductCategoryPage({
           </ProductCardSection>
         </div>
       </div>
-      <ProductAddButton />
+      <ProductCreateButton />
     </>
   )
 }
