+++ src/widgets/product/common/category-dropdown/CategoryDropdown.stories.tsx
@@ -1,9 +1,9 @@
+import { useEffect } from 'react'
 import { useForm } from 'react-hook-form'
 import { Meta, StoryObj } from '@storybook/react'
 import { ProductPayload } from '@shared/types'
 import CategoryDropdown from './CategoryDropdown'
-import { defaultValues } from '@widgets/product/product-home/components/product-add-modal/form/defaultValues'
-import { useEffect } from 'react'
+import { defaultValues } from '@widgets/product/product-home/components/product-create-modal/form/defaultValues'
 
 function CategoryDropdownWrapper() {
   const {
