+++ src/widgets/product/product-home/components/index.ts
@@ -1,9 +1,9 @@
 export { default as ProductHot } from './product-hot/ProductHot'
 export { default as ProductRating } from './product-rating/ProductRating'
-export { default as ProductAddButton } from './product-add-button/ProductAddButton'
+export { default as ProductCreateButton } from './product-create-button/ProductCreateButton'
 export { default as ProductSearched } from './product-searched'
 export { default as ProductEmpty } from './product-empty/ProductEmpty'
 
-export * from './product-add-modal'
+export * from './product-create-modal'
 export * from './ranking'
 export * from './category-menu'
