--- src/widgets/product/product-home/components/product-add-modal/index.ts
@@ -1,2 +1,0 @@
-export { default as ProductAddModal } from './modal/ProductAddModal'
-export { default as ProductAddForm } from './form/ProductAddForm'
