+++ src/widgets/product/product-home/components/product-create-modal/form/defaultValues.ts
