+++ src/widgets/product/product-home/components/product-create-modal/index.ts
@@ -1,0 +1,2 @@
+export { default as ProductCreateModal } from './modal/ProductCreateModal'
+export { default as ProductCreateForm } from './form/ProductCreateForm'
